RAC-31: refactor: use right directory, improve readability and add phpdoc
