+++ src/Akeneo/Pim/Enrichment/Component/Product/Normalizer/Versioning/EntityWithQuantifiedAssociations/QuantifiedAssociationsNormalizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product;
+namespace Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\EntityWithQuantifiedAssociations;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithQuantifiedAssociationsInterface;
 use Symfony\Component\Serializer\Normalizer\CacheableSupportsMethodInterface;
@@ -10,8 +10,6 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class QuantifiedAssociationsNormalizer implements NormalizerInterface, CacheableSupportsMethodInterface
 {
-    protected $supportedFormats = ['flat'];
-
     /**
      * {@inheritdoc}
      *
@@ -23,14 +21,14 @@ class QuantifiedAssociationsNormalizer implements NormalizerInterface, Cacheable
 
         $results = [];
         foreach ($quantifiedAssociationsNormalized as $associationTypeCode => $linkTypes) {
-            foreach ($linkTypes['products'] as $productLink) {
-                $flatAssociationKey = $associationTypeCode . '-products-' . $productLink['identifier'];
-                $results[$flatAssociationKey] = $productLink['quantity'];
+            foreach ($linkTypes['products'] as $quantifiedLink) {
+                $flatAssociationKey = $associationTypeCode . '-products-' . $quantifiedLink['identifier'];
+                $results[$flatAssociationKey] = $quantifiedLink['quantity'];
             }
 
-            foreach ($linkTypes['product_models'] as $productModelLink) {
-                $flatAssociationKey = $associationTypeCode . '-product_models-' . $productModelLink['identifier'];
-                $results[$flatAssociationKey] = $productModelLink['quantity'];
+            foreach ($linkTypes['product_models'] as $quantifiedLink) {
+                $flatAssociationKey = $associationTypeCode . '-product_models-' . $quantifiedLink['identifier'];
+                $results[$flatAssociationKey] = $quantifiedLink['quantity'];
             }
         }
 
@@ -42,7 +40,7 @@ class QuantifiedAssociationsNormalizer implements NormalizerInterface, Cacheable
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof EntityWithQuantifiedAssociationsInterface && in_array($format, $this->supportedFormats);
+        return $data instanceof EntityWithQuantifiedAssociationsInterface && $format === 'flat';
     }
 
     public function hasCacheableSupportsMethod(): bool
