+++ tests/back/Pim/Enrichment/Specification/Component/Product/Normalizer/Versioning/EntityWithQuantifiedAssociations/QuantifiedAssociationsNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product;
+namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\EntityWithQuantifiedAssociations;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\GroupInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
